+++ src/game/player/elements/image/Image.tsx
@@ -7,6 +7,10 @@ import {useRatio} from "@player/provider/ratio";
 import clsx from "clsx";
 import {useDisplayable} from "@player/elements/displayable/Displayable";
 import {Utils} from "@core/common/Utils";
+import {ITransition} from "@lib/game/nlcore/elements/transition/type";
+import {useTransition} from "@player/lib/useTransition";
+import {useGame} from "@player/provider/game-state";
+import {deepMerge} from "@lib/util/data";
 
 /**@internal */
 export default function Image(
@@ -20,7 +24,7 @@ export default function Image(
     const {ratio} = useRatio();
     const [wearables, setWearables] = useState<GameImage[]>([]);
     const ref = useRef<HTMLImageElement>(null);
-    const {ref: transformRef, transitionProps} = useDisplayable<HTMLImageElement>({
+    const {ref: transformRef, transition} = useDisplayable({
         element: image,
         state: image.transformState,
         skipTransform: state.game.config.elements.img.allowSkipTransform,
@@ -30,15 +34,6 @@ export default function Image(
                 width: "100%",
                 height: "100%",
             } : {}),
-        },
-        transitionProp: {
-            src: GameImage.getSrcURL(image) || GameImage.DefaultImagePlaceholder,
-            style: {
-                ...(state.game.config.app.debug ? {
-                    outline: "1px solid red",
-                } : {}),
-                transformOrigin: "center",
-            },
         },
     });
 
@@ -74,23 +69,12 @@ export default function Image(
             ref={transformRef}
             className={"absolute"}
         >
-            {transitionProps.map((elementProps, index) => (
-                <AspectScaleImage
-                    key={index}
-                    props={{
-                        className: "absolute",
-                        style: {
-                            display: "block",
-                            position: "unset",
-                        },
-                        ...elementProps,
-                    }}
-                    id={elementProps.src}
-                    ref={index === 0 ? ref : undefined}
-                    onWidthChange={handleWidthChange}
-                    autoFit={image.config.autoFit}
-                />
-            ))}
+            <ImageTransition
+                transition={transition}
+                ref={ref}
+                onWidthChange={handleWidthChange}
+                image={image}
+            />
             <div
                 className={clsx("w-full h-full top-0 left-0 absolute")}
             >
@@ -106,3 +90,66 @@ export default function Image(
         </Inspect.mDiv>
     );
 };
+
+function ImageTransition(
+    {
+        transition,
+        ref,
+        onWidthChange,
+        image,
+    }: {
+        transition: ITransition | undefined,
+        ref: React.RefObject<HTMLImageElement | null>,
+        onWidthChange: () => void,
+        image: GameImage;
+    }
+) {
+    const {game} = useGame();
+    const [transitionProps] = useTransition<HTMLImageElement>({
+        transition,
+        props: {
+            src: GameImage.getSrcURL(image) || GameImage.DefaultImagePlaceholder,
+            style: {
+                ...(game.config.app.debug ? {
+                    outline: "1px solid red",
+                } : {}),
+                transformOrigin: "center",
+            },
+        }
+    });
+
+    const elementStyles = [
+        {
+            style: {
+                position: "absolute",
+            }
+        },
+        {
+            style: {
+                maxWidth: "none",
+                maxHeight: "none",
+                transform: "translate(-50%, -50%)",
+                top: "50%",
+                left: "50%",
+                position: "absolute",
+            }
+        }
+    ];
+
+    return (
+        <>
+            <div className={"relative h-full w-full"}>
+            {transitionProps.map((elementProps, index) => {
+                return <AspectScaleImage
+                    key={index}
+                    props={deepMerge(elementProps, elementStyles[index] || {})}
+                    id={elementProps.src as any}
+                    ref={index === 0 ? ref : undefined}
+                    onWidthChange={onWidthChange}
+                    autoFit={image.config.autoFit}
+                />;
+            })}
+            </div>
+        </>
+    );
+}